@@ -9,10 +9,14 @@
         }
     ],
     "require": {
-        "php": "^7.1"
+        "php": "^8.1",
+        "typo3/cms-core": "^11.5 || ^12.4",
+        "ext-json": "*"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.1"
+        "phpspec/prophecy": "^1.17",
+        "phpspec/prophecy-phpunit": "^2.0",
+        "phpunit/phpunit": "^9.6.9"
     },
     "autoload": {
         "psr-4": {
@@ -30,5 +34,10 @@
     },
     "suggest": {
         "ext-readline": "For better CLI support"
+    },
+    "extra": {
+        "typo3/cms": {
+            "extension-key": "document_storage"
+        }
     }
 }
