@@ -1,11 +1,11 @@
 <?php
+
 declare(strict_types=1);
 
 namespace Cundd\DocumentStorage;
 
 interface Constants
 {
-
     /**
      * Name of the property that holds the data
      */
