@@ -1,4 +1,5 @@
 <?php
+
 declare(strict_types=1);
 
 namespace Cundd\DocumentStorage\Domain\Model;
@@ -10,22 +11,22 @@ use ArrayAccess;
  */
 class Dictionary extends Document implements ArrayAccess
 {
-    public function offsetExists($offset)
+    public function offsetExists($offset): bool
     {
         return (bool)$this->valueForKey((string)$offset);
     }
 
-    public function offsetGet($offset)
+    public function offsetGet($offset): mixed
     {
         return $this->valueForKey((string)$offset);
     }
 
-    public function offsetSet($offset, $value)
+    public function offsetSet($offset, $value): void
     {
         $this->setValueForKey((string)$offset, $value);
     }
 
-    public function offsetUnset($offset)
+    public function offsetUnset(mixed $offset): void
     {
         $this->setValueForKey((string)$offset, null);
     }
